+++ src/gt4py/next/program_processors/runners/dace/__init__.py
@@ -7,7 +7,7 @@
 # SPDX-License-Identifier: BSD-3-Clause
 
 
-from gt4py.next.program_processors.runners.dace.gtir_sdfg import build_sdfg_from_gtir
+from gt4py.next.program_processors.runners.dace.gtir_lower import build_sdfg_from_gtir
 from gt4py.next.program_processors.runners.dace.sdfg_callable import get_sdfg_args
 from gt4py.next.program_processors.runners.dace.workflow.backend import (
     make_dace_backend,
