+++ src/gt4py/next/program_processors/runners/dace/gtir_dataflow.py
@@ -34,9 +34,9 @@ from gt4py.next.iterator import builtins as gtir_builtins, ir as gtir
 from gt4py.next.iterator.ir_utils import common_pattern_matcher as cpm, ir_makers as im
 from gt4py.next.iterator.transforms import symbol_ref_utils
 from gt4py.next.program_processors.runners.dace import (
+    gtir_lower,
+    gtir_lower_utils,
     gtir_python_codegen,
-    gtir_sdfg,
-    gtir_sdfg_utils,
     utils as gtx_dace_utils,
 )
 from gt4py.next.type_system import type_info as ti, type_specifications as ts
@@ -360,7 +360,7 @@ class LambdaToDataflow(eve.NodeVisitor):
 
     sdfg: dace.SDFG
     state: dace.SDFGState
-    subgraph_builder: gtir_sdfg.DataflowBuilder
+    subgraph_builder: gtir_lower.DataflowBuilder
     input_edges: list[DataflowInputEdge] = dataclasses.field(default_factory=lambda: [])
     symbol_map: dict[
         str,
@@ -732,7 +732,7 @@ class LambdaToDataflow(eve.NodeVisitor):
             if isinstance(arg, tuple):
                 ptype = get_tuple_type(arg)  # type: ignore[arg-type]
                 psymbol = im.sym(pname, ptype)
-                psymbol_tree = gtir_sdfg_utils.make_symbol_tree(pname, ptype)
+                psymbol_tree = gtir_lower_utils.make_symbol_tree(pname, ptype)
                 deref_on_input_memlet = pname in direct_deref_iterators
                 inner_arg = gtx_utils.tree_map(
                     lambda tsym,
@@ -839,7 +839,7 @@ class LambdaToDataflow(eve.NodeVisitor):
         )
 
         nsdfg = dace.SDFG(self.unique_nsdfg_name(prefix="if_stmt"))
-        nsdfg.debuginfo = gtir_sdfg_utils.debug_info(node, default=self.sdfg.debuginfo)
+        nsdfg.debuginfo = gtir_lower_utils.debug_info(node, default=self.sdfg.debuginfo)
 
         # create states inside the nested SDFG for the if-branches
         if_region = dace.sdfg.state.ConditionalBlock("if")
@@ -904,7 +904,7 @@ class LambdaToDataflow(eve.NodeVisitor):
                 edge.connect(map_entry=None)
 
             if isinstance(node.type, ts.TupleType):
-                out_symbol_tree = gtir_sdfg_utils.make_symbol_tree("__output", node.type)
+                out_symbol_tree = gtir_lower_utils.make_symbol_tree("__output", node.type)
                 outer_value = gtx_utils.tree_map(
                     lambda x, y, nstate=nstate: self._visit_if_branch_result(nsdfg, nstate, x, y)
                 )(output_tree, out_symbol_tree)
@@ -1022,7 +1022,7 @@ class LambdaToDataflow(eve.NodeVisitor):
         )
         neighbors_node = self.state.add_access(neighbors_temp)
         offset_type = gtx_common.Dimension(offset, gtx_common.DimensionKind.LOCAL)
-        neighbor_idx = gtir_sdfg_utils.get_map_variable(offset_type)
+        neighbor_idx = gtir_lower_utils.get_map_variable(offset_type)
 
         index_connector = "__index"
         output_connector = "__val"
@@ -1177,7 +1177,7 @@ class LambdaToDataflow(eve.NodeVisitor):
             raise ValueError("Unexpected arguments to map expression with different neighborhood.")
         offset_type, offset_provider_type = next(iter(input_connectivity_types.items()))
         local_size = offset_provider_type.max_neighbors
-        map_index = gtir_sdfg_utils.get_map_variable(offset_type)
+        map_index = gtir_lower_utils.get_map_variable(offset_type)
 
         # The dataflow we build in this class has some loose connections on input edges.
         # These edges are described as set of nodes, that will have to be connected to
@@ -1215,7 +1215,7 @@ class LambdaToDataflow(eve.NodeVisitor):
             connectivity_desc = self.sdfg.arrays[connectivity]
             connectivity_desc.transient = False
 
-            origin_map_index = gtir_sdfg_utils.get_map_variable(offset_provider_type.source_dim)
+            origin_map_index = gtir_lower_utils.get_map_variable(offset_provider_type.source_dim)
 
             connectivity_slice = self._construct_local_view(
                 MemletExpr(
@@ -1283,7 +1283,7 @@ class LambdaToDataflow(eve.NodeVisitor):
         corresponding neighbor index in the connectivity table is valid, or the
         identity value if the neighbor index is missing.
         """
-        origin_map_index = gtir_sdfg_utils.get_map_variable(offset_provider_type.source_dim)
+        origin_map_index = gtir_lower_utils.get_map_variable(offset_provider_type.source_dim)
 
         assert (
             isinstance(input_expr.gt_dtype, ts.ListType)
@@ -1864,7 +1864,7 @@ class LambdaToDataflow(eve.NodeVisitor):
 def translate_lambda_to_dataflow(
     sdfg: dace.SDFG,
     state: dace.SDFGState,
-    sdfg_builder: gtir_sdfg.DataflowBuilder,
+    sdfg_builder: gtir_lower.DataflowBuilder,
     node: gtir.Lambda,
     args: Sequence[
         IteratorExpr
