+++ src/gt4py/next/program_processors/runners/dace/gtir_lower_primitives.py
@@ -9,25 +9,21 @@
 from __future__ import annotations
 
 import abc
-import dataclasses
-from typing import TYPE_CHECKING, Any, Final, Iterable, Optional, Protocol, Sequence, TypeAlias
+from typing import TYPE_CHECKING, Any, Iterable, Optional, Protocol
 
 import dace
 from dace import subsets as dace_subsets
 
 from gt4py.next import common as gtx_common, utils as gtx_utils
-from gt4py.next.ffront import fbuiltins as gtx_fbuiltins
 from gt4py.next.iterator import ir as gtir
-from gt4py.next.iterator.ir_utils import (
-    common_pattern_matcher as cpm,
-    domain_utils,
-    ir_makers as im,
-)
+from gt4py.next.iterator.ir_utils import common_pattern_matcher as cpm, ir_makers as im
 from gt4py.next.program_processors.runners.dace import (
     gtir_dataflow,
+    gtir_domain,
+    gtir_lower,
+    gtir_lower_types,
+    gtir_lower_utils,
     gtir_python_codegen,
-    gtir_sdfg,
-    gtir_sdfg_utils,
     utils as gtx_dace_utils,
 )
 from gt4py.next.program_processors.runners.dace.gtir_lower_scan import translate_scan
@@ -35,247 +31,7 @@ from gt4py.next.type_system import type_info as ti, type_specifications as ts
 
 
 if TYPE_CHECKING:
-    from gt4py.next.program_processors.runners.dace import gtir_sdfg
-
-
-def get_domain_indices(
-    dims: Sequence[gtx_common.Dimension], origin: Optional[Sequence[dace.symbolic.SymExpr]]
-) -> dace_subsets.Indices:
-    """
-    Helper function to construct the list of indices for a field domain, applying
-    an optional origin in each dimension as start index.
-
-    Args:
-        dims: The field dimensions.
-        origin: The domain start index in each dimension. If set to `None`, assume all zeros.
-
-    Returns:
-        A list of indices for field access in dace arrays. As this list is returned
-        as `dace.subsets.Indices`, it should be converted to `dace.subsets.Range` before
-        being used in memlet subset because ranges are better supported throughout DaCe.
-    """
-    assert len(dims) != 0
-    index_variables = [
-        dace.symbolic.pystr_to_symbolic(gtir_sdfg_utils.get_map_variable(dim)) for dim in dims
-    ]
-    origin = [0] * len(index_variables) if origin is None else origin
-    return dace_subsets.Indices(
-        [index - start_index for index, start_index in zip(index_variables, origin, strict=True)]
-    )
-
-
-@dataclasses.dataclass(frozen=True)
-class FieldopData:
-    """
-    Abstraction to represent data (scalars, arrays) during the lowering to SDFG.
-
-    Attribute 'local_offset' must always be set for `FieldType` data with a local
-    dimension generated from neighbors access in unstructured domain, and indicates
-    the name of the offset provider used to generate the list of neighbor values.
-
-    Args:
-        dc_node: DaCe access node to the data storage.
-        gt_type: GT4Py type definition, which includes the field domain information.
-        origin: Tuple of start indices, in each dimension, for `FieldType` data.
-            Pass an empty tuple for `ScalarType` data or zero-dimensional fields.
-    """
-
-    dc_node: dace.nodes.AccessNode
-    gt_type: ts.FieldType | ts.ScalarType
-    origin: tuple[dace.symbolic.SymbolicType, ...]
-
-    def __post_init__(self) -> None:
-        """Implements a sanity check on the constructed data type."""
-        assert (
-            len(self.origin) == 0
-            if isinstance(self.gt_type, ts.ScalarType)
-            else len(self.origin) == len(self.gt_type.dims)
-        )
-
-    def map_to_parent_sdfg(
-        self,
-        sdfg_builder: gtir_sdfg.SDFGBuilder,
-        inner_sdfg: dace.SDFG,
-        outer_sdfg: dace.SDFG,
-        outer_sdfg_state: dace.SDFGState,
-        symbol_mapping: dict[str, dace.symbolic.SymbolicType],
-    ) -> FieldopData:
-        """
-        Make the data descriptor which 'self' refers to, and which is located inside
-        a NestedSDFG, available in its parent SDFG.
-
-        Thus, it turns 'self' into a non-transient array and creates a new data
-        descriptor inside the parent SDFG, with same shape and strides.
-        """
-        inner_desc = self.dc_node.desc(inner_sdfg)
-        assert inner_desc.transient
-        inner_desc.transient = False
-
-        if isinstance(self.gt_type, ts.ScalarType):
-            outer, outer_desc = sdfg_builder.add_temp_scalar(outer_sdfg, inner_desc.dtype)
-            outer_origin = []
-        else:
-            outer, outer_desc = sdfg_builder.add_temp_array_like(outer_sdfg, inner_desc)
-            # We cannot use a copy of the inner data descriptor directly, we have to apply the symbol mapping.
-            dace.symbolic.safe_replace(
-                symbol_mapping,
-                lambda m: dace.sdfg.replace_properties_dict(outer_desc, m),
-            )
-            # Same applies to the symbols used as field origin (the domain range start)
-            outer_origin = [
-                gtx_dace_utils.safe_replace_symbolic(val, symbol_mapping) for val in self.origin
-            ]
-
-        outer_node = outer_sdfg_state.add_access(outer)
-        return FieldopData(outer_node, self.gt_type, tuple(outer_origin))
-
-    def get_local_view(
-        self, domain: FieldopDomain, sdfg: dace.SDFG
-    ) -> gtir_dataflow.IteratorExpr | gtir_dataflow.MemletExpr:
-        """Helper method to access a field in local view, given the compute domain of a field operator."""
-        if isinstance(self.gt_type, ts.ScalarType):
-            assert isinstance(self.dc_node.desc(sdfg), dace.data.Scalar)
-            return gtir_dataflow.MemletExpr(
-                dc_node=self.dc_node,
-                gt_dtype=self.gt_type,
-                subset=dace_subsets.Range.from_string("0"),
-            )
-
-        if isinstance(self.gt_type, ts.FieldType):
-            it_indices: dict[gtx_common.Dimension, gtir_dataflow.DataExpr]
-            if isinstance(self.dc_node.desc(sdfg), dace.data.Scalar):
-                assert len(self.gt_type.dims) == 0  # zero-dimensional field
-                it_indices = {}
-            else:
-                # The invariant below is ensured by calling `make_field()` to construct `FieldopData`.
-                # The `make_field` constructor converts any local dimension, if present, to `ListType`
-                # element type, while leaving the field domain with all global dimensions.
-                assert all(dim != gtx_common.DimensionKind.LOCAL for dim in self.gt_type.dims)
-                domain_dims = [dim for dim, _, _ in domain]
-                domain_indices = get_domain_indices(domain_dims, origin=None)
-                it_indices = {
-                    dim: gtir_dataflow.SymbolExpr(index, INDEX_DTYPE)
-                    for dim, index in zip(domain_dims, domain_indices)
-                }
-            field_origin = [
-                (dim, dace.symbolic.SymExpr(0) if self.origin is None else self.origin[i])
-                for i, dim in enumerate(self.gt_type.dims)
-            ]
-            return gtir_dataflow.IteratorExpr(
-                self.dc_node, self.gt_type.dtype, field_origin, it_indices
-            )
-
-        raise NotImplementedError(f"Node type {type(self.gt_type)} not supported.")
-
-    def get_symbol_mapping(
-        self, dataname: str, sdfg: dace.SDFG
-    ) -> dict[str, dace.symbolic.SymExpr]:
-        """
-        Helper method to create the symbol mapping for array storage in a nested SDFG.
-
-        Args:
-            dataname: Name of the data container insiode the nested SDFG.
-            sdfg: The parent SDFG where the `FieldopData` object lives.
-
-        Returns:
-            Mapping from symbols in nested SDFG to the corresponding symbolic values
-            in the parent SDFG. This includes the range start and stop symbols (used
-            to calculate the array shape as range 'stop - start') and the strides.
-        """
-        if isinstance(self.gt_type, ts.ScalarType):
-            return {}
-        ndims = len(self.gt_type.dims)
-        outer_desc = self.dc_node.desc(sdfg)
-        assert isinstance(outer_desc, dace.data.Array)
-        # origin and size of the local dimension, in case of a field with `ListType` data,
-        # are assumed to be compiled-time values (not symbolic), therefore the start and
-        # stop range symbols of the inner field only extend over the global dimensions
-        return (
-            {gtx_dace_utils.range_start_symbol(dataname, i): (self.origin[i]) for i in range(ndims)}
-            | {
-                gtx_dace_utils.range_stop_symbol(dataname, i): (
-                    self.origin[i] + outer_desc.shape[i]
-                )
-                for i in range(ndims)
-            }
-            | {
-                gtx_dace_utils.field_stride_symbol_name(dataname, i): stride
-                for i, stride in enumerate(outer_desc.strides)
-            }
-        )
-
-
-FieldopDomain: TypeAlias = list[
-    tuple[gtx_common.Dimension, dace.symbolic.SymbolicType, dace.symbolic.SymbolicType]
-]
-"""
-Domain of a field operator represented as a list of tuples with 3 elements:
- - dimension definition
- - symbolic expression for lower bound (inclusive)
- - symbolic expression for upper bound (exclusive)
-"""
-
-
-FieldopResult: TypeAlias = FieldopData | tuple[FieldopData | tuple, ...]
-"""Result of a field operator, can be either a field or a tuple fields."""
-
-
-INDEX_DTYPE: Final[dace.typeclass] = dace.dtype_to_typeclass(gtx_fbuiltins.IndexType)
-"""Data type used for field indexing."""
-
-
-def get_arg_symbol_mapping(
-    dataname: str, arg: FieldopResult, sdfg: dace.SDFG
-) -> dict[str, dace.symbolic.SymExpr]:
-    """
-    Helper method to build the mapping from inner to outer SDFG of all symbols
-    used for storage of a field or a tuple of fields.
-
-    Args:
-        dataname: The storage name inside the nested SDFG.
-        arg: The argument field in the parent SDFG.
-        sdfg: The parent SDFG where the argument field lives.
-
-    Returns:
-        A mapping from inner symbol names to values or symbolic definitions
-        in the parent SDFG.
-    """
-    if isinstance(arg, FieldopData):
-        return arg.get_symbol_mapping(dataname, sdfg)
-
-    symbol_mapping: dict[str, dace.symbolic.SymExpr] = {}
-    for i, elem in enumerate(arg):
-        dataname_elem = f"{dataname}_{i}"
-        symbol_mapping |= get_arg_symbol_mapping(dataname_elem, elem, sdfg)
-
-    return symbol_mapping
-
-
-def get_tuple_type(data: tuple[FieldopResult, ...]) -> ts.TupleType:
-    """
-    Compute the `ts.TupleType` corresponding to the tuple structure of `FieldopResult`.
-    """
-    return ts.TupleType(
-        types=[get_tuple_type(d) if isinstance(d, tuple) else d.gt_type for d in data]
-    )
-
-
-def flatten_tuples(name: str, arg: FieldopResult) -> list[tuple[str, FieldopData]]:
-    """
-    Visit a `FieldopResult`, potentially containing nested tuples, and construct a list
-    of pairs `(str, FieldopData)` containing the symbol name of each tuple field and
-    the corresponding `FieldopData`.
-    """
-    if isinstance(arg, tuple):
-        tuple_type = get_tuple_type(arg)
-        tuple_symbols = gtir_sdfg_utils.flatten_tuple_fields(name, tuple_type)
-        tuple_data_fields = gtx_utils.flatten_nested_tuple(arg)
-        return [
-            (str(tsym.id), tfield)
-            for tsym, tfield in zip(tuple_symbols, tuple_data_fields, strict=True)
-        ]
-    else:
-        return [(name, arg)]
+    from gt4py.next.program_processors.runners.dace import gtir_lower
 
 
 class PrimitiveTranslator(Protocol):
@@ -285,8 +41,8 @@ class PrimitiveTranslator(Protocol):
         node: gtir.Node,
         sdfg: dace.SDFG,
         state: dace.SDFGState,
-        sdfg_builder: gtir_sdfg.SDFGBuilder,
-    ) -> FieldopResult:
+        sdfg_builder: gtir_lower.SDFGBuilder,
+    ) -> gtir_lower_types.FieldopResult:
         """Creates the dataflow subgraph representing a GTIR primitive function.
 
         This method is used by derived classes to build a specialized subgraph
@@ -310,8 +66,8 @@ def _parse_fieldop_arg(
     node: gtir.Expr,
     sdfg: dace.SDFG,
     state: dace.SDFGState,
-    sdfg_builder: gtir_sdfg.SDFGBuilder,
-    domain: FieldopDomain,
+    sdfg_builder: gtir_lower.SDFGBuilder,
+    domain: gtir_domain.FieldopDomain,
 ) -> (
     gtir_dataflow.IteratorExpr
     | gtir_dataflow.MemletExpr
@@ -321,53 +77,22 @@ def _parse_fieldop_arg(
 
     arg = sdfg_builder.visit(node, sdfg=sdfg, head_state=state)
 
-    if isinstance(arg, FieldopData):
+    if isinstance(arg, gtir_lower_types.FieldopData):
         return arg.get_local_view(domain, sdfg)
     else:
         # handle tuples of fields
         return gtx_utils.tree_map(lambda targ: targ.get_local_view(domain))(arg)
 
 
-def get_field_layout(
-    domain: FieldopDomain,
-) -> tuple[list[gtx_common.Dimension], list[dace.symbolic.SymExpr], list[dace.symbolic.SymExpr]]:
-    """
-    Parse the field operator domain and generates the shape of the result field.
-
-    It should be enough to allocate an array with shape (upper_bound - lower_bound)
-    but this would require to use array offset for compensate for the start index.
-    Suppose that a field operator executes on domain [2,N-2], the dace array to store
-    the result only needs size (N-4), but this would require to compensate all array
-    accesses with offset -2 (which corresponds to -lower_bound). Instead, we choose
-    to allocate (N-2), leaving positions [0:2] unused. The reason is that array offset
-    is known to cause issues to SDFG inlining. Besides, map fusion will in any case
-    eliminate most of transient arrays.
-
-    Args:
-        domain: The field operator domain.
-
-    Returns:
-        A tuple of three lists containing:
-            - the domain dimensions
-            - the domain origin, that is the start indices in all dimensions
-            - the domain size in each dimension
-    """
-    if len(domain) == 0:
-        return [], [], []
-    domain_dims, domain_lbs, domain_ubs = zip(*domain)
-    domain_sizes = [(ub - lb) for lb, ub in zip(domain_lbs, domain_ubs)]
-    return list(domain_dims), list(domain_lbs), domain_sizes
-
-
 def _create_field_operator_impl(
-    sdfg_builder: gtir_sdfg.SDFGBuilder,
+    sdfg_builder: gtir_lower.SDFGBuilder,
     sdfg: dace.SDFG,
     state: dace.SDFGState,
-    domain: FieldopDomain,
+    domain: gtir_domain.FieldopDomain,
     output_edge: gtir_dataflow.DataflowOutputEdge,
     output_type: ts.FieldType,
     map_exit: dace.nodes.MapExit,
-) -> FieldopData:
+) -> gtir_lower_types.FieldopData:
     """
     Helper method to allocate a temporary array that stores one field computed
     by a field operator.
@@ -390,13 +115,13 @@ def _create_field_operator_impl(
     dataflow_output_desc = output_edge.result.dc_node.desc(sdfg)
 
     # the memory layout of the output field follows the field operator compute domain
-    field_dims, field_origin, field_shape = get_field_layout(domain)
+    field_dims, field_origin, field_shape = gtir_domain.get_field_layout(domain)
     if len(domain) == 0:
         # The field operator computes a zero-dimensional field, and the data subset
         # is set later depending on the element type (`ts.ListType` or `ts.ScalarType`)
         field_subset = dace_subsets.Range([])
     else:
-        field_indices = get_domain_indices(field_dims, field_origin)
+        field_indices = gtir_domain.get_domain_indices(field_dims, field_origin)
         field_subset = dace_subsets.Range.from_indices(field_indices)
 
     if isinstance(output_edge.result.gt_dtype, ts.ScalarType):
@@ -430,7 +155,7 @@ def _create_field_operator_impl(
     # and here the edge writing the dataflow result data through the map exit node
     output_edge.connect(map_exit, field_node, field_subset)
 
-    return FieldopData(
+    return gtir_lower_types.FieldopData(
         field_node, ts.FieldType(field_dims, output_edge.result.gt_dtype), tuple(field_origin)
     )
 
@@ -438,12 +163,12 @@ def _create_field_operator_impl(
 def _create_field_operator(
     sdfg: dace.SDFG,
     state: dace.SDFGState,
-    domain: FieldopDomain,
+    domain: gtir_domain.FieldopDomain,
     node_type: ts.FieldType | ts.TupleType,
-    sdfg_builder: gtir_sdfg.SDFGBuilder,
+    sdfg_builder: gtir_lower.SDFGBuilder,
     input_edges: Iterable[gtir_dataflow.DataflowInputEdge],
     output_tree: tuple[gtir_dataflow.DataflowOutputEdge | tuple[Any, ...], ...],
-) -> FieldopResult:
+) -> gtir_lower_types.FieldopResult:
     """
     Helper method to build the output of a field operator, which can consist of
     a single field or a tuple of fields.
@@ -473,7 +198,7 @@ def _create_field_operator(
     else:
         # create map range corresponding to the field operator domain
         map_range = {
-            gtir_sdfg_utils.get_map_variable(dim): f"{lower_bound}:{upper_bound}"
+            gtir_lower_utils.get_map_variable(dim): f"{lower_bound}:{upper_bound}"
             for dim, lower_bound, upper_bound in domain
         }
     map_entry, map_exit = sdfg_builder.add_map("fieldop", state, map_range)
@@ -492,7 +217,7 @@ def _create_field_operator(
         )
     else:
         # handle tuples of fields
-        output_symbol_tree = gtir_sdfg_utils.make_symbol_tree("x", node_type)
+        output_symbol_tree = gtir_lower_utils.make_symbol_tree("x", node_type)
         return gtx_utils.tree_map(
             lambda output_edge, output_sym: _create_field_operator_impl(
                 sdfg_builder, sdfg, state, domain, output_edge, output_sym.type, map_exit
@@ -500,49 +225,12 @@ def _create_field_operator(
         )(output_tree, output_symbol_tree)
 
 
-def extract_domain(node: gtir.Node) -> FieldopDomain:
-    """
-    Visits the domain of a field operator and returns a list of dimensions and
-    the corresponding lower and upper bounds. The returned lower bound is inclusive,
-    the upper bound is exclusive: [lower_bound, upper_bound[
-    """
-
-    domain = []
-
-    if cpm.is_call_to(node, ("cartesian_domain", "unstructured_domain")):
-        for named_range in node.args:
-            assert cpm.is_call_to(named_range, "named_range")
-            assert len(named_range.args) == 3
-            axis = named_range.args[0]
-            assert isinstance(axis, gtir.AxisLiteral)
-            lower_bound, upper_bound = (
-                gtir_sdfg_utils.get_symbolic(arg) for arg in named_range.args[1:3]
-            )
-            dim = gtx_common.Dimension(axis.value, axis.kind)
-            domain.append((dim, lower_bound, upper_bound))
-
-    elif isinstance(node, domain_utils.SymbolicDomain):
-        for dim, drange in node.ranges.items():
-            domain.append(
-                (
-                    dim,
-                    gtir_sdfg_utils.get_symbolic(drange.start),
-                    gtir_sdfg_utils.get_symbolic(drange.stop),
-                )
-            )
-
-    else:
-        raise ValueError(f"Invalid domain {node}.")
-
-    return domain
-
-
 def translate_as_fieldop(
     node: gtir.Node,
     sdfg: dace.SDFG,
     state: dace.SDFGState,
-    sdfg_builder: gtir_sdfg.SDFGBuilder,
-) -> FieldopResult:
+    sdfg_builder: gtir_lower.SDFGBuilder,
+) -> gtir_lower_types.FieldopResult:
     """
     Generates the dataflow subgraph for the `as_fieldop` builtin function.
 
@@ -584,7 +272,7 @@ def translate_as_fieldop(
         )
 
     # parse the domain of the field operator
-    domain = extract_domain(domain_expr)
+    domain = gtir_domain.extract_domain(domain_expr)
 
     # visit the list of arguments to be passed to the lambda expression
     fieldop_args = [_parse_fieldop_arg(arg, sdfg, state, sdfg_builder, domain) for arg in node.args]
@@ -602,12 +290,12 @@ def translate_as_fieldop(
 def _construct_if_branch_output(
     sdfg: dace.SDFG,
     state: dace.SDFGState,
-    sdfg_builder: gtir_sdfg.SDFGBuilder,
+    sdfg_builder: gtir_lower.SDFGBuilder,
     domain: gtir.Expr,
     sym: gtir.Sym,
-    true_br: FieldopData,
-    false_br: FieldopData,
-) -> FieldopData:
+    true_br: gtir_lower_types.FieldopData,
+    false_br: gtir_lower_types.FieldopData,
+) -> gtir_lower_types.FieldopData:
     """
     Helper function called by `translate_if()` to allocate a temporary field to store
     the result of an if expression.
@@ -620,11 +308,11 @@ def _construct_if_branch_output(
         dtype = gtx_dace_utils.as_dace_type(sym.type)
         out, _ = sdfg_builder.add_temp_scalar(sdfg, dtype)
         out_node = state.add_access(out)
-        return FieldopData(out_node, sym.type, origin=())
+        return gtir_lower_types.FieldopData(out_node, sym.type, origin=())
 
     assert isinstance(out_type, ts.FieldType)
     assert isinstance(sym.type, ts.FieldType)
-    dims, origin, shape = get_field_layout(extract_domain(domain))
+    dims, origin, shape = gtir_domain.get_field_layout(gtir_domain.extract_domain(domain))
     assert dims == out_type.dims
 
     if isinstance(out_type.dtype, ts.ScalarType):
@@ -643,14 +331,14 @@ def _construct_if_branch_output(
     out, _ = sdfg_builder.add_temp_array(sdfg, shape, dtype)
     out_node = state.add_access(out)
 
-    return FieldopData(out_node, out_type, tuple(origin))
+    return gtir_lower_types.FieldopData(out_node, out_type, tuple(origin))
 
 
 def _write_if_branch_output(
     sdfg: dace.SDFG,
     state: dace.SDFGState,
-    src: FieldopData,
-    dst: FieldopData,
+    src: gtir_lower_types.FieldopData,
+    dst: gtir_lower_types.FieldopData,
 ) -> None:
     """
     Helper function called by `translate_if()` to write the result of an if-branch,
@@ -702,8 +390,8 @@ def translate_if(
     node: gtir.Node,
     sdfg: dace.SDFG,
     state: dace.SDFGState,
-    sdfg_builder: gtir_sdfg.SDFGBuilder,
-) -> FieldopResult:
+    sdfg_builder: gtir_lower.SDFGBuilder,
+) -> gtir_lower_types.FieldopResult:
     """Generates the dataflow subgraph for the `if_` builtin function."""
     assert cpm.is_call_to(node, "if_")
     assert len(node.args) == 3
@@ -752,7 +440,7 @@ def translate_if(
     )
 
     if isinstance(node.type, ts.TupleType):
-        symbol_tree = gtir_sdfg_utils.make_symbol_tree("x", node.type)
+        symbol_tree = gtir_lower_utils.make_symbol_tree("x", node.type)
         if isinstance(node.annex.domain, tuple):
             domain_tree = node.annex.domain
         else:
@@ -807,8 +495,8 @@ def translate_index(
     node: gtir.Node,
     sdfg: dace.SDFG,
     state: dace.SDFGState,
-    sdfg_builder: gtir_sdfg.SDFGBuilder,
-) -> FieldopResult:
+    sdfg_builder: gtir_lower.SDFGBuilder,
+) -> gtir_lower_types.FieldopResult:
     """
     Lowers the `index` builtin function to a mapped tasklet that writes the dimension
     index values to a transient array. The extent of the index range is taken from
@@ -818,16 +506,16 @@ def translate_index(
     assert isinstance(node.type, ts.FieldType)
 
     assert "domain" in node.annex
-    domain = extract_domain(node.annex.domain)
+    domain = gtir_domain.extract_domain(node.annex.domain)
     assert len(domain) == 1
     dim, _, _ = domain[0]
-    dim_index = gtir_sdfg_utils.get_map_variable(dim)
+    dim_index = gtir_lower_utils.get_map_variable(dim)
 
-    index_data, _ = sdfg_builder.add_temp_scalar(sdfg, INDEX_DTYPE)
+    index_data, _ = sdfg_builder.add_temp_scalar(sdfg, gtir_lower_types.INDEX_DTYPE)
     index_node = state.add_access(index_data)
     index_value = gtir_dataflow.ValueExpr(
         dc_node=index_node,
-        gt_dtype=gtx_dace_utils.as_itir_type(INDEX_DTYPE),
+        gt_dtype=gtx_dace_utils.as_itir_type(gtir_lower_types.INDEX_DTYPE),
     )
     index_write_tasklet = sdfg_builder.add_tasklet(
         "index",
@@ -856,10 +544,10 @@ def translate_index(
 def _get_data_nodes(
     sdfg: dace.SDFG,
     state: dace.SDFGState,
-    sdfg_builder: gtir_sdfg.SDFGBuilder,
+    sdfg_builder: gtir_lower.SDFGBuilder,
     data_name: str,
     data_type: ts.DataType,
-) -> FieldopResult:
+) -> gtir_lower_types.FieldopResult:
     if isinstance(data_type, ts.FieldType):
         data_node = state.add_access(data_name)
         return sdfg_builder.make_field(data_node, data_type)
@@ -871,10 +559,10 @@ def _get_data_nodes(
             )
         else:
             data_node = state.add_access(data_name)
-        return FieldopData(data_node, data_type, origin=())
+        return gtir_lower_types.FieldopData(data_node, data_type, origin=())
 
     elif isinstance(data_type, ts.TupleType):
-        symbol_tree = gtir_sdfg_utils.make_symbol_tree(data_name, data_type)
+        symbol_tree = gtir_lower_utils.make_symbol_tree(data_name, data_type)
         return gtx_utils.tree_map(
             lambda sym: _get_data_nodes(sdfg, state, sdfg_builder, sym.id, sym.type)
         )(symbol_tree)
@@ -886,7 +574,7 @@ def _get_data_nodes(
 def _get_symbolic_value(
     sdfg: dace.SDFG,
     state: dace.SDFGState,
-    sdfg_builder: gtir_sdfg.SDFGBuilder,
+    sdfg_builder: gtir_lower.SDFGBuilder,
     symbolic_expr: dace.symbolic.SymExpr,
     scalar_type: ts.ScalarType,
     temp_name: Optional[str] = None,
@@ -919,23 +607,23 @@ def translate_literal(
     node: gtir.Node,
     sdfg: dace.SDFG,
     state: dace.SDFGState,
-    sdfg_builder: gtir_sdfg.SDFGBuilder,
-) -> FieldopResult:
+    sdfg_builder: gtir_lower.SDFGBuilder,
+) -> gtir_lower_types.FieldopResult:
     """Generates the dataflow subgraph for a `ir.Literal` node."""
     assert isinstance(node, gtir.Literal)
 
     data_type = node.type
     data_node = _get_symbolic_value(sdfg, state, sdfg_builder, node.value, data_type)
 
-    return FieldopData(data_node, data_type, origin=())
+    return gtir_lower_types.FieldopData(data_node, data_type, origin=())
 
 
 def translate_make_tuple(
     node: gtir.Node,
     sdfg: dace.SDFG,
     state: dace.SDFGState,
-    sdfg_builder: gtir_sdfg.SDFGBuilder,
-) -> FieldopResult:
+    sdfg_builder: gtir_lower.SDFGBuilder,
+) -> gtir_lower_types.FieldopResult:
     assert cpm.is_call_to(node, "make_tuple")
     return tuple(
         sdfg_builder.visit(
@@ -951,8 +639,8 @@ def translate_tuple_get(
     node: gtir.Node,
     sdfg: dace.SDFG,
     state: dace.SDFGState,
-    sdfg_builder: gtir_sdfg.SDFGBuilder,
-) -> FieldopResult:
+    sdfg_builder: gtir_lower.SDFGBuilder,
+) -> gtir_lower_types.FieldopResult:
     assert cpm.is_call_to(node, "tuple_get")
     assert len(node.args) == 2
 
@@ -966,9 +654,9 @@ def translate_tuple_get(
         sdfg=sdfg,
         head_state=state,
     )
-    if isinstance(data_nodes, FieldopData):
+    if isinstance(data_nodes, gtir_lower_types.FieldopData):
         raise ValueError(f"Invalid tuple expression {node}")
-    unused_arg_nodes: Iterable[FieldopData] = gtx_utils.flatten_nested_tuple(
+    unused_arg_nodes: Iterable[gtir_lower_types.FieldopData] = gtx_utils.flatten_nested_tuple(
         tuple(arg for i, arg in enumerate(data_nodes) if i != index)
     )
     state.remove_nodes_from(
@@ -981,8 +669,8 @@ def translate_scalar_expr(
     node: gtir.Node,
     sdfg: dace.SDFG,
     state: dace.SDFGState,
-    sdfg_builder: gtir_sdfg.SDFGBuilder,
-) -> FieldopResult:
+    sdfg_builder: gtir_lower.SDFGBuilder,
+) -> gtir_lower_types.FieldopResult:
     assert isinstance(node, gtir.FunCall)
     assert isinstance(node.type, ts.ScalarType)
 
@@ -1009,7 +697,10 @@ def translate_scalar_expr(
                 sdfg=sdfg,
                 head_state=state,
             )
-            if not (isinstance(arg, FieldopData) and isinstance(node.type, ts.ScalarType)):
+            if not (
+                isinstance(arg, gtir_lower_types.FieldopData)
+                and isinstance(node.type, ts.ScalarType)
+            ):
                 raise ValueError(f"Invalid argument to scalar expression {arg_expr}.")
             param = f"__arg{i}"
             args.append(arg.dc_node)
@@ -1049,15 +740,15 @@ def translate_scalar_expr(
         dace.Memlet(data=temp_name, subset="0"),
     )
 
-    return FieldopData(temp_node, node.type, origin=())
+    return gtir_lower_types.FieldopData(temp_node, node.type, origin=())
 
 
 def translate_symbol_ref(
     node: gtir.Node,
     sdfg: dace.SDFG,
     state: dace.SDFGState,
-    sdfg_builder: gtir_sdfg.SDFGBuilder,
-) -> FieldopResult:
+    sdfg_builder: gtir_lower.SDFGBuilder,
+) -> gtir_lower_types.FieldopResult:
     """Generates the dataflow subgraph for a `ir.SymRef` node."""
     assert isinstance(node, gtir.SymRef)
