+++ src/gt4py/next/program_processors/runners/dace/gtir_lower_utils.py
@@ -16,7 +16,7 @@ from gt4py import eve
 from gt4py.next import common as gtx_common, utils as gtx_utils
 from gt4py.next.iterator import ir as gtir
 from gt4py.next.iterator.ir_utils import ir_makers as im
-from gt4py.next.program_processors.runners.dace import gtir_python_codegen
+from gt4py.next.program_processors.runners.dace import gtir_lower_types, gtir_python_codegen
 from gt4py.next.type_system import type_specifications as ts
 
 
@@ -34,6 +34,33 @@ def debug_info(
             filename=location.filename,
         )
     return default
+
+
+def get_arg_symbol_mapping(
+    dataname: str, arg: gtir_lower_types.FieldopResult, sdfg: dace.SDFG
+) -> dict[str, dace.symbolic.SymExpr]:
+    """
+    Helper method to build the mapping from inner to outer SDFG of all symbols
+    used for storage of a field or a tuple of fields.
+
+    Args:
+        dataname: The storage name inside the nested SDFG.
+        arg: The argument field in the parent SDFG.
+        sdfg: The parent SDFG where the argument field lives.
+
+    Returns:
+        A mapping from inner symbol names to values or symbolic definitions
+        in the parent SDFG.
+    """
+    if isinstance(arg, gtir_lower_types.FieldopData):
+        return arg.get_symbol_mapping(dataname, sdfg)
+
+    symbol_mapping: dict[str, dace.symbolic.SymExpr] = {}
+    for i, elem in enumerate(arg):
+        dataname_elem = f"{dataname}_{i}"
+        symbol_mapping |= get_arg_symbol_mapping(dataname_elem, elem, sdfg)
+
+    return symbol_mapping
 
 
 def get_map_variable(dim: gtx_common.Dimension) -> str:
