+++ src/gt4py/next/program_processors/runners/dace/transformations/loop_blocking.py
@@ -20,7 +20,7 @@ from dace.sdfg import graph as dace_graph, nodes as dace_nodes
 from dace.transformation import helpers as dace_helpers
 
 from gt4py.next import common as gtx_common
-from gt4py.next.program_processors.runners.dace import gtir_sdfg_utils
+from gt4py.next.program_processors.runners.dace import gtir_lower_utils
 
 
 @dace_properties.make_properties
@@ -84,7 +84,7 @@ class LoopBlocking(dace_transformation.SingleStateTransformation):
     ) -> None:
         super().__init__()
         if isinstance(blocking_parameter, gtx_common.Dimension):
-            blocking_parameter = gtir_sdfg_utils.get_map_variable(blocking_parameter)
+            blocking_parameter = gtir_lower_utils.get_map_variable(blocking_parameter)
         if blocking_parameter is not None:
             self.blocking_parameter = blocking_parameter
         if blocking_size is not None:
